+++ CHANGELOG.md
@@ -1,8 +1,26 @@
-﻿## [2.5.9] ##
+## [2.5.10] ##
 ### Modified AddOns ###
+  * Aurora 11.2.7.2
   * nibRealUI
+  * RealUI_Inventory
 
 ### Changed ###
+  * chg: moved from NUM_CHAT_WINDOWS to Constants.ChatFrameConstants.MaxChatWindows
+
+## Detailed Changes ##
+[2.5.9]: https://github.com/RealUI/RealUI/compare/2.5.9...2.5.10
+
+## [2.5.9] ##
+### Modified AddOns ###
+  * Aurora 11.2.7.1
+  * nibRealUI
+  * nibRealUI_Config
+  * RealUI_Inventory
+
+### Changed ###
+  * fix: bug in CUSTOM_CLASS_COLORS/LOCALIZED_CLASS_NAMES_MALE https://github.com/Stanzilla/WoWUIBugs/issues/798
+  * fix: new enum.FrameTutorialAccount replaces LE_FRAME_TUTORIAL_ACCOUNT
+  * fix: duplicate guid
   * add: housing xp tracking added
 
 ## Detailed Changes ##